Lazy-load /search and /search/test; pull SearchPage out of main bundle
The 20kB-gzipped main-bundle bump on the previous PR was mostly the
diagnostic tooling — ScoreBreakdown, EntityResultRow, the whole
SearchPage component. None of it needs to ship until the user
actually navigates to /search.

Following the existing chats.tsx + chats.*.lazy.tsx pattern in the
repo:
- src/routes/_user/-search-page.tsx (NEW) — co-located non-route
  component (`-` prefix per project convention) holding SearchPage +
  every sub-component (PhraseResultRow, EntityResultRow, ScoreBreakdown,
  TypeBadge, FilterPill, EmptyResults, SystemMessage, WelcomeState,
  QuickFiltersPanel).
- src/routes/_user/search.tsx — trimmed to ~20 lines: just the route
  metadata (validateSearch, beforeLoad) and an Outlet. This is the
  only search-related file in the eager bundle.
- src/routes/_user/search.index.lazy.tsx (NEW) — replaces
  search.index.tsx. createLazyFileRoute. Imports SearchPage.
- src/routes/_user/search.test.lazy.tsx (NEW) — replaces
  search.test.tsx. createLazyFileRoute. Imports SearchPage with
  diagnostic prop.

Build chunks land as expected:
- search-*.js — 199 bytes (eager layout)
- -search-page-*.js — 22 kB (the heavy component, lazy)
- search.index.lazy-*.js — 909 bytes (lazy route shell)
- search.test.lazy-*.js — 926 bytes (lazy route shell)
- use-merged-search-*.js — 9 kB (lazy)
- use-semantic-search-*.js — 1.8 kB (lazy)

Total search code stays roughly the same (~35 kB across chunks), but
none of it ships in the main bundle anymore. The diagnostic title
override on /search/test went away as a small UX trim — both routes
show "Search" in the title bar; the URL distinguishes them. Saves
needing a paired metadata file alongside each .lazy.tsx.
