+++ src/routes/_user/search.index.lazy.tsx
@@ -1,0 +1,6 @@
+import { createLazyFileRoute } from '@tanstack/react-router'
+import { SearchPage } from './-search-page'
+
+export const Route = createLazyFileRoute('/_user/search/')({
+	component: () => <SearchPage />,
+})
