--- src/routes/_user/search.index.tsx
@@ -1,6 +1,0 @@
-import { createFileRoute } from '@tanstack/react-router'
-import { SearchPage } from './search'
-
-export const Route = createFileRoute('/_user/search/')({
-	component: () => <SearchPage />,
-})
